Refactor jobV2
