+++ src/dashboard/src/pages/JobV2/Tabs/Metrics.tsx
@@ -8,7 +8,7 @@ import {
   CardMedia
 } from '@material-ui/core';
 
-import Context from './Context';
+import Context from '../Context';
 
 const Metrics: FunctionComponent = () => {
   const { cluster, job } = useContext(Context);
