+++ src/dashboard/src/pages/JobV2/useRouteParams.ts
@@ -1,0 +1,10 @@
+import { useParams } from 'react-router';
+
+interface RouteParams {
+  clusterId: string;
+  jobId: string;
+}
+
+const useRouteParams = () => useParams<RouteParams>();
+
+export default useRouteParams;
